update (#4)
* Warn if GraphViz not installed

Only lets you choose DOT as an output format, and gives a message warning that GraphViz is not installed so options are limited.
As client side generation is added, this should still be suitable.

* Updated Dutch translation

* Create SECURITY.md

* Update README.md

* Update README.md

* Add Russian translation

Thanks to POEditor.com user "oleg"

* Update README.md

* Change "Preview" back to "Update"

This change reverts change of button from "Update" to "Preview" based on discussion in #80

* Rename "Cousins" and enforce

Change "Cousins" to "All relatives". If "All relatives" checked, ensure "Siblings" is also checked. If "Siblings" unchecked while "All relatives" checked, then uncheck "All relatives"

* Allow override of cart

This adds the ability to override the cart contents, so settings can be used to select records instead of using the contents of the cart, even if there are items in the cart.

* Fix blank nodesep/ranksep

If spacing values are blank or otherwise not numbers, set to default 100%

Resolves #89

* Update SECURITY.md

* Create CONTRIBUTING.md

* Update CONTRIBUTING.md

* Update README.md

* Update CONTRIBUTING.md

* Update CONTRIBUTING.md

* Update translations

* Attempt to fix issue with numeric value

* Respect XREF setting in Vesta

When using the Vesta suite of modules, you can enable an option to show the XREF in the person's name. This means the option to show XREF no longer works.
In this commit, we make a change so that the GVExport option will successfully disable this, but we also change it so that if the option to show XREF in name is enabled in Vesta (or another similar module), then the option to display the XREF on individuals is turned on by default (otherwise it is off by default).

* Add "Anybody" relation option (#86)

* Add "Anybody" relation option

This option will include any linked person in the selected generations (respecting max ancestors/descendants settings). In practice, this adds the relatives of spouses.

* Respect disabling of ancestors/descendants options

* Add parent of unrelated child

Update to fix parent of unrelated child not showing when "Anyone" option set

* Correctly trace tree for "Anyone" option

This update brings a change to how the tree is traversed to support the "Anyone" option, that will traverse anyone know matter how distantly related.

* Fix unrelated child's missing parents

* Change non-blood related colours

Making them slightly more vivid so they can be seen better

* Don't enable "anyone" option by default.

This change disables the "Anyone" option by default.

* Fix partner or partner not included

Small change to include spouses of spouses if "Anyone" option selected

* Don't override parents value

With "Anyone" option there is a lot more rechecking of records. If family record already has "has parents" set to true, don't override it and set it to false just because of the path it was found the second time.
Resolves issue with combined mode sometimes not showing a family and instead just showing the family ID.

* Treat cart as empty if no indi (thanks @hartenthaler)

Adds changes from #70 and also moves most javascript functions to separate file.

* Render in browser on page load (#103)

Change default ancestor and descendant levels to 2 to reduce records, then enable loading the browser rendering as soon as page is loaded. Data from cookie used instead of defaults if cookie is set.

* Update clippings cart terms (#105)

Update the terms used when there are items in the clippings cart

* Underline preferred name (#106)

Search for "starredname" class in full name, replacing this with <U></U> to underline.

* Remove "reset" parameter from URL (#107)

On page load, removes the "reset" parameter from the URL.
Resolves  104

* Fix adopted records being wrong colour (#108)

When an individual is adopted, they are not a blood relative (unless also related in a way other than by adoption). This change fixes the colouring of adopted records by updating existing non-functioning code to work with recent versions of webtrees, and then turning this into a function to be called from various places affected.
Resolves #36

* Fix related colours when starting from adopted individual (#111)

* Fix related colours when starting from adopted individual

If adopted person was chosen as starting individual, siblings were assumed as related and this spread through the tree. Fixed by adding check for if the individual being added is not related to their own family, if they aren't related then set $rel to false for this round of recursion.

* Update translations (#112)

* Fix clippings cart options (#113)

When items in clippings cart, the option to ignore was not working as when branches had been merged some of the old code was merged in. This resolves the issue.

* Fix cart (#114)

* Fix cart issue

Fix issue with UI when items in cart

* Fix cart (#115)

Currently on each page load the URL is updated to remove the "reset" parameter. We should only push a new URL ot the history if there actually was a "reset" parameter set

Co-authored-by: Neriderc <89974636+Neriderc@users.noreply.github.com>
Co-authored-by: TheDutchJewel <46184729+TheDutchJewel@users.noreply.github.com>